// helpers/types.ts
export enum SupportTokens {
    ETHER = 'ETHER',
    DAI = 'DAI',
    AAVE = 'AAVE',
    TUSD = 'TUSD',
    BAT = 'BAT',
    WETH = 'WETH',
    USDC = 'USDC',
    USDT = 'USDT',
    SUSD = 'SUSD',
    ZRX = 'ZRX',
    MKR = 'MKR',
    WBTC = 'WBTC',
    LINK = 'LINK',
    KNC = 'KNC',
    MANA = 'MANA',
    REN = 'REN',
    SNX = 'SNX',
    BUSD = 'BUSD',
    USD = 'USD',
    YFI = 'YFI',
    UNI = 'UNI',
    ENJ = 'ENJ'
}

export enum LoanType {
    NORMAL = 0,
    MARGIN_SWAP_PROTOCOL = 1,
    MINNING_SWAP_PROTOCOL = 2,
    INVALID_PROTOCOL = 3
}